chore: extract 'findApi' and 'createError' into separate modules
