+++ packages/dashboard-backend/src/devworkspace-client/services/helpers/createError.ts
@@ -10,26 +10,7 @@
  *   Red Hat, Inc. - initial API and implementation
  */
 
-import * as k8s from '@kubernetes/client-node';
 import { helpers } from '@eclipse-che/common';
-
-export async function findApi(
-  apisApi: k8s.ApisApi,
-  apiName: string,
-  version?: string,
-): Promise<boolean> {
-  const resp = await apisApi.getAPIVersions();
-  const groups = resp.body.groups;
-  return groups.some((apiGroup: k8s.V1APIGroup) => {
-    if (version) {
-      return (
-        apiGroup.name === apiName &&
-        apiGroup.versions.some(versionGroup => versionGroup.version === version)
-      );
-    }
-    return apiGroup.name === apiName;
-  });
-}
 
 class DevWorkspaceClientError extends Error {
   statusCode: number;
